+++ src/providers/lua-types.ts
@@ -1,11 +1,6 @@
 /*---------------------------------------------------------------------------------------------
  *  Copyright (c) RStudio, PBC. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
- 
-Error running filter /Users/jjallaire/quarto/quarto-ext/code-filename/_extensions/code-filename/code-filename.lua:
-...ode-filename/_extensions/code-filename/code-filename.lua:59: ')' expected (to close '(' at line 54) near 'end'
-
- 
  *--------------------------------------------------------------------------------------------*/
 
 import * as path from "path";