+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.39.0 (Unreleased)
 
 - Provide completions and diagnostics for profile specific \_quarto.yml
+- Cleanup residual Quarto Preview panel on startup/reload
 
 ## 1.38.0 (Release on 10 September 2022)